Extract inline components and hooks from page files
Move 5 hooks from session/[id]/page.tsx to src/hooks/:
- useRefetchOnReconnect, useSessionState, useSessionMessages,
  useWorkCompleteNotification, useClaudeState

Extract 3 components from new/page.tsx to src/components/:
- RepoSelector, BranchSelector, IssueSelector

Create shared useInfiniteScroll hook to deduplicate the
IntersectionObserver pattern used by RepoSelector and IssueSelector.

Session page reduced from 636 to 238 lines, new session page from
540 to 192 lines.

Closes #198
