+++ src/app/new/page.tsx
@@ -1,6 +1,6 @@
 'use client';
 
-import { useState, useEffect, useCallback, useRef } from 'react';
+import { useState, useCallback } from 'react';
 import { useRouter } from 'next/navigation';
 import Link from 'next/link';
 import { AuthGuard } from '@/components/AuthGuard';
@@ -11,360 +11,12 @@ import { Input } from '@/components/ui/input';
 import { Label } from '@/components/ui/label';
 import { Card, CardContent, CardHeader, CardTitle } from '@/components/ui/card';
 import { Alert, AlertDescription } from '@/components/ui/alert';
-import {
-  Select,
-  SelectContent,
-  SelectItem,
-  SelectTrigger,
-  SelectValue,
-} from '@/components/ui/select';
 import { Spinner } from '@/components/ui/spinner';
-import { Badge } from '@/components/ui/badge';
-import { cn } from '@/lib/utils';
+import { RepoSelector } from '@/components/RepoSelector';
+import { BranchSelector } from '@/components/BranchSelector';
+import { IssueSelector } from '@/components/IssueSelector';
+import type { Repo } from '@/components/RepoSelector';
 import type { Issue } from '@/lib/types';
-import { Star } from 'lucide-react';
-
-interface Repo {
-  id: number;
-  fullName: string;
-  name: string;
-  owner: string;
-  description: string | null;
-  private: boolean;
-  defaultBranch: string;
-}
-
-function RepoSelector({
-  selectedRepo,
-  onSelect,
-}: {
-  selectedRepo: Repo | null;
-  onSelect: (repo: Repo) => void;
-}) {
-  const [search, setSearch] = useState('');
-  const [debouncedSearch, setDebouncedSearch] = useState('');
-
-  useEffect(() => {
-    const timer = setTimeout(() => setDebouncedSearch(search), 300);
-    return () => clearTimeout(timer);
-  }, [search]);
-
-  const { data, isLoading, fetchNextPage, hasNextPage, isFetchingNextPage } =
-    trpc.github.listRepos.useInfiniteQuery(
-      { search: debouncedSearch || undefined, perPage: 20 },
-      {
-        getNextPageParam: (lastPage) => lastPage.nextCursor,
-      }
-    );
-
-  // Fetch favorites to sort repos and show star icons
-  const { data: favoritesData } = trpc.repoSettings.listFavorites.useQuery();
-  const favorites = new Set(favoritesData?.favorites ?? []);
-
-  const toggleFavorite = trpc.repoSettings.toggleFavorite.useMutation();
-  const utils = trpc.useUtils();
-
-  const handleToggleFavorite = (e: React.MouseEvent, repoFullName: string) => {
-    e.stopPropagation(); // Don't trigger repo selection
-    const newIsFavorite = !favorites.has(repoFullName);
-    toggleFavorite.mutate(
-      { repoFullName, isFavorite: newIsFavorite },
-      {
-        onSuccess: () => {
-          utils.repoSettings.listFavorites.invalidate();
-        },
-      }
-    );
-  };
-
-  const rawRepos = data?.pages.flatMap((p) => p.repos) || [];
-
-  // Sort repos with favorites first
-  const repos = [...rawRepos].sort((a, b) => {
-    const aFav = favorites.has(a.fullName);
-    const bFav = favorites.has(b.fullName);
-    if (aFav && !bFav) return -1;
-    if (!aFav && bFav) return 1;
-    return 0;
-  });
-
-  const scrollRef = useRef<HTMLDivElement>(null);
-  const sentinelRef = useRef<HTMLLIElement>(null);
-
-  useEffect(() => {
-    const sentinel = sentinelRef.current;
-    const scrollContainer = scrollRef.current;
-    if (!sentinel || !scrollContainer) return;
-
-    const observer = new IntersectionObserver(
-      (entries) => {
-        if (entries[0].isIntersecting && hasNextPage && !isFetchingNextPage) {
-          fetchNextPage();
-        }
-      },
-      { root: scrollContainer, rootMargin: '100px' }
-    );
-
-    observer.observe(sentinel);
-    return () => observer.disconnect();
-  }, [hasNextPage, isFetchingNextPage, fetchNextPage]);
-
-  return (
-    <div className="space-y-4">
-      <div className="space-y-2">
-        <Label>Search repositories</Label>
-        <Input
-          type="text"
-          value={search}
-          onChange={(e) => setSearch(e.target.value)}
-          placeholder="Search your repositories..."
-        />
-      </div>
-
-      <div ref={scrollRef} className="border rounded-lg max-h-64 overflow-y-auto">
-        {isLoading ? (
-          <div className="flex justify-center py-8">
-            <Spinner />
-          </div>
-        ) : repos.length === 0 ? (
-          <div className="text-center py-8 text-muted-foreground">No repositories found</div>
-        ) : (
-          <ul className="divide-y divide-border">
-            {repos.map((repo) => {
-              const isFavorite = favorites.has(repo.fullName);
-              return (
-                <li
-                  key={repo.id}
-                  onClick={() => onSelect(repo)}
-                  className={cn(
-                    'px-4 py-3 cursor-pointer hover:bg-muted/50 transition-colors',
-                    selectedRepo?.id === repo.id && 'bg-primary/10'
-                  )}
-                >
-                  <div className="flex items-center justify-between gap-2">
-                    <button
-                      type="button"
-                      onClick={(e) => handleToggleFavorite(e, repo.fullName)}
-                      className="p-1 hover:bg-muted rounded transition-colors shrink-0"
-                      title={isFavorite ? 'Remove from favorites' : 'Add to favorites'}
-                    >
-                      <Star
-                        className={cn(
-                          'h-4 w-4',
-                          isFavorite ? 'text-yellow-500 fill-yellow-500' : 'text-muted-foreground'
-                        )}
-                      />
-                    </button>
-                    <div className="flex-1 min-w-0">
-                      <p className="text-sm font-medium">{repo.fullName}</p>
-                      {repo.description && (
-                        <p className="text-xs text-muted-foreground truncate max-w-md">
-                          {repo.description}
-                        </p>
-                      )}
-                    </div>
-                    {repo.private && (
-                      <span className="text-xs text-muted-foreground shrink-0">Private</span>
-                    )}
-                  </div>
-                </li>
-              );
-            })}
-            {hasNextPage && (
-              <li ref={sentinelRef} className="px-4 py-3 flex justify-center">
-                <Spinner size="sm" />
-              </li>
-            )}
-          </ul>
-        )}
-      </div>
-    </div>
-  );
-}
-
-function BranchSelector({
-  repoFullName,
-  selectedBranch,
-  onSelect,
-}: {
-  repoFullName: string;
-  selectedBranch: string;
-  onSelect: (branch: string) => void;
-}) {
-  const { data, isLoading } = trpc.github.listBranches.useQuery(
-    { repoFullName },
-    { enabled: !!repoFullName }
-  );
-
-  const handleSelect = useCallback(
-    (branch: string) => {
-      onSelect(branch);
-    },
-    [onSelect]
-  );
-
-  useEffect(() => {
-    if (data?.defaultBranch && !selectedBranch) {
-      handleSelect(data.defaultBranch);
-    }
-  }, [data, selectedBranch, handleSelect]);
-
-  if (isLoading) {
-    return (
-      <div className="flex items-center gap-2 text-muted-foreground">
-        <Spinner size="sm" />
-        <span>Loading branches...</span>
-      </div>
-    );
-  }
-
-  const branches = data?.branches || [];
-
-  return (
-    <div className="space-y-2">
-      <Label>Branch</Label>
-      <Select value={selectedBranch} onValueChange={onSelect}>
-        <SelectTrigger>
-          <SelectValue placeholder="Select a branch" />
-        </SelectTrigger>
-        <SelectContent>
-          {branches.map((branch) => (
-            <SelectItem key={branch.name} value={branch.name}>
-              {branch.name}
-              {branch.name === data?.defaultBranch ? ' (default)' : ''}
-            </SelectItem>
-          ))}
-        </SelectContent>
-      </Select>
-    </div>
-  );
-}
-
-function IssueSelector({
-  repoFullName,
-  selectedIssue,
-  onSelect,
-}: {
-  repoFullName: string;
-  selectedIssue: Issue | null;
-  onSelect: (issue: Issue | null) => void;
-}) {
-  const [search, setSearch] = useState('');
-  const [debouncedSearch, setDebouncedSearch] = useState('');
-
-  useEffect(() => {
-    const timer = setTimeout(() => setDebouncedSearch(search), 300);
-    return () => clearTimeout(timer);
-  }, [search]);
-
-  const { data, isLoading, fetchNextPage, hasNextPage, isFetchingNextPage } =
-    trpc.github.listIssues.useInfiniteQuery(
-      { repoFullName, search: debouncedSearch || undefined, perPage: 15 },
-      {
-        getNextPageParam: (lastPage) => lastPage.nextCursor,
-        enabled: !!repoFullName,
-      }
-    );
-
-  const issues = data?.pages.flatMap((p) => p.issues) || [];
-
-  const issueScrollRef = useRef<HTMLDivElement>(null);
-  const issueSentinelRef = useRef<HTMLLIElement>(null);
-
-  useEffect(() => {
-    const sentinel = issueSentinelRef.current;
-    const scrollContainer = issueScrollRef.current;
-    if (!sentinel || !scrollContainer) return;
-
-    const observer = new IntersectionObserver(
-      (entries) => {
-        if (entries[0].isIntersecting && hasNextPage && !isFetchingNextPage) {
-          fetchNextPage();
-        }
-      },
-      { root: scrollContainer, rootMargin: '100px' }
-    );
-
-    observer.observe(sentinel);
-    return () => observer.disconnect();
-  }, [hasNextPage, isFetchingNextPage, fetchNextPage]);
-
-  return (
-    <div className="space-y-4">
-      <div className="space-y-2">
-        <Label>Link to GitHub issue (optional)</Label>
-        <Input
-          type="text"
-          value={search}
-          onChange={(e) => setSearch(e.target.value)}
-          placeholder="Search issues..."
-        />
-      </div>
-
-      <div ref={issueScrollRef} className="border rounded-lg max-h-48 overflow-y-auto">
-        {isLoading ? (
-          <div className="flex justify-center py-6">
-            <Spinner />
-          </div>
-        ) : issues.length === 0 ? (
-          <div className="text-center py-6 text-muted-foreground text-sm">
-            {search ? 'No issues found' : 'No open issues'}
-          </div>
-        ) : (
-          <ul className="divide-y divide-border">
-            {selectedIssue && (
-              <li
-                onClick={() => onSelect(null)}
-                className="px-4 py-2 cursor-pointer hover:bg-muted/50 transition-colors bg-muted/30"
-              >
-                <div className="flex items-center gap-2 text-sm text-muted-foreground">
-                  <span>Clear selection</span>
-                </div>
-              </li>
-            )}
-            {issues.map((issue) => (
-              <li
-                key={issue.id}
-                onClick={() => onSelect(issue)}
-                className={cn(
-                  'px-4 py-2 cursor-pointer hover:bg-muted/50 transition-colors',
-                  selectedIssue?.id === issue.id && 'bg-primary/10'
-                )}
-              >
-                <div className="flex items-start gap-2">
-                  <span className="text-xs text-muted-foreground shrink-0">#{issue.number}</span>
-                  <div className="flex-1 min-w-0">
-                    <p className="text-sm font-medium truncate">{issue.title}</p>
-                    <div className="flex flex-wrap gap-1 mt-1">
-                      {issue.labels.slice(0, 3).map((label) => (
-                        <Badge
-                          key={label.name}
-                          variant="outline"
-                          className="text-xs py-0"
-                          style={{
-                            borderColor: `#${label.color}`,
-                            color: `#${label.color}`,
-                          }}
-                        >
-                          {label.name}
-                        </Badge>
-                      ))}
-                    </div>
-                  </div>
-                </div>
-              </li>
-            ))}
-            {hasNextPage && (
-              <li ref={issueSentinelRef} className="px-4 py-2 flex justify-center">
-                <Spinner size="sm" />
-              </li>
-            )}
-          </ul>
-        )}
-      </div>
-    </div>
-  );
-}
 
 function generateIssuePrompt(issue: Issue, repoFullName: string): string {
   const issueUrl = `https://github.com/${repoFullName}/issues/${issue.number}`;
