+++ src/app/session/[id]/page.tsx
@@ -1,6 +1,6 @@
 'use client';
 
-import { useCallback, useMemo, useEffect, useRef, use } from 'react';
+import { useCallback, useEffect, useRef, use } from 'react';
 import Link from 'next/link';
 import { AuthGuard } from '@/components/AuthGuard';
 import { Header } from '@/components/Header';
@@ -8,413 +8,15 @@ import { SessionHeader } from '@/components/SessionHeader';
 import { MessageList } from '@/components/MessageList';
 import { PromptInput } from '@/components/PromptInput';
 import { ClaudeStatusIndicator } from '@/components/ClaudeStatusIndicator';
-import { trpc } from '@/lib/trpc';
 import { Button } from '@/components/ui/button';
 import { Spinner } from '@/components/ui/spinner';
 import { useNotification } from '@/hooks/useNotification';
 import { useWorkingIndicator } from '@/hooks/useWorkingIndicator';
 import { useWorkingContext } from '@/lib/working-context';
-
-interface Message {
-  id: string;
-  type: string;
-  content: unknown;
-  sequence: number;
-}
-
-const MESSAGE_PAGE_SIZE = 20;
-
-/**
- * Hook to refetch data when the app regains visibility or network reconnects.
- * This handles cases where SSE connection was lost and the UI shows stale state.
- */
-function useRefetchOnReconnect(refetch: () => void) {
-  useEffect(() => {
-    const handleVisibilityChange = () => {
-      if (document.visibilityState === 'visible') {
-        refetch();
-      }
-    };
-
-    const handleOnline = () => {
-      refetch();
-    };
-
-    document.addEventListener('visibilitychange', handleVisibilityChange);
-    window.addEventListener('online', handleOnline);
-
-    return () => {
-      document.removeEventListener('visibilitychange', handleVisibilityChange);
-      window.removeEventListener('online', handleOnline);
-    };
-  }, [refetch]);
-}
-
-/**
- * Hook for managing session state: fetching session data, SSE updates, and start/stop mutations.
- */
-function useSessionState(sessionId: string) {
-  const utils = trpc.useUtils();
-
-  // Fetch session details
-  const { data: sessionData, isLoading, refetch } = trpc.sessions.get.useQuery({ sessionId });
-
-  // Refetch session data when app regains visibility or network reconnects
-  useRefetchOnReconnect(refetch);
-
-  // Subscribe to session updates via SSE - update cache directly
-  trpc.sse.onSessionUpdate.useSubscription(
-    { sessionId },
-    {
-      onData: (trackedData) => {
-        const session = trackedData.data.session as NonNullable<typeof sessionData>['session'];
-        utils.sessions.get.setData({ sessionId }, { session });
-      },
-      onError: (err) => {
-        console.error('Session SSE error:', err);
-      },
-    }
-  );
-
-  // Mutations - update cache directly from returned data
-  const startMutation = trpc.sessions.start.useMutation({
-    onSuccess: (data) => {
-      utils.sessions.get.setData({ sessionId }, { session: data.session });
-    },
-  });
-
-  const stopMutation = trpc.sessions.stop.useMutation({
-    onSuccess: (data) => {
-      utils.sessions.get.setData({ sessionId }, { session: data.session });
-    },
-  });
-
-  // The API endpoint is "delete" but it now archives instead of permanently deleting
-  // Session update comes via SSE subscription (lines 64-75), so no onSuccess handler needed
-  const archiveMutation = trpc.sessions.delete.useMutation();
-
-  const start = useCallback(() => {
-    startMutation.mutate({ sessionId });
-  }, [sessionId, startMutation]);
-
-  const stop = useCallback(() => {
-    stopMutation.mutate({ sessionId });
-  }, [sessionId, stopMutation]);
-
-  const archive = useCallback(() => {
-    archiveMutation.mutate({ sessionId });
-  }, [sessionId, archiveMutation]);
-
-  return {
-    session: sessionData?.session,
-    isLoading,
-    start,
-    stop,
-    archive,
-    isStarting: startMutation.isPending,
-    isStopping: stopMutation.isPending,
-    isArchiving: archiveMutation.isPending,
-  };
-}
-
-/**
- * Hook for managing message state: history pagination, SSE updates for new messages, and token usage.
- */
-function useSessionMessages(sessionId: string) {
-  const utils = trpc.useUtils();
-
-  // Bidirectional infinite query for message history
-  // - fetchNextPage: loads older messages (backward) when user scrolls up
-  // - New messages arrive via SSE and are added directly to the cache
-  const {
-    data: historyData,
-    isLoading,
-    isFetchingNextPage,
-    hasNextPage,
-    fetchNextPage,
-  } = trpc.claude.getHistory.useInfiniteQuery(
-    { sessionId, limit: MESSAGE_PAGE_SIZE },
-    {
-      // Limit stored pages to prevent memory growth
-      // With MESSAGE_PAGE_SIZE messages per page, this keeps up to 10000 messages in memory
-      maxPages: 500,
-      // Message data is immutable - never refetch automatically
-      staleTime: Infinity,
-      initialCursor: {
-        direction: 'backward',
-        sequence: undefined,
-      },
-      // For loading OLDER messages (user scrolls up)
-      getNextPageParam: (lastPage, allPages) => {
-        if (!lastPage.hasMore) return undefined;
-        // Find the oldest sequence across ALL pages (not just lastPage, which might be empty)
-        let oldestSequence: number | undefined;
-        for (const page of allPages) {
-          for (const msg of page.messages) {
-            if (oldestSequence === undefined || msg.sequence < oldestSequence) {
-              oldestSequence = msg.sequence;
-            }
-          }
-        }
-        return { sequence: oldestSequence, direction: 'backward' as const };
-      },
-    }
-  );
-
-  // Fetch token usage stats (computed server-side from all messages)
-  const { data: tokenUsageData, refetch: refetchTokenUsage } = trpc.claude.getTokenUsage.useQuery(
-    { sessionId },
-    {
-      // Refetch less frequently since it's just for display
-      refetchOnWindowFocus: false,
-    }
-  );
-
-  // Compute newest sequence from cache for SSE catch-up cursor
-  const newestSequence = useMemo(() => {
-    if (!historyData?.pages) return undefined;
-    let newest: number | undefined;
-    for (const page of historyData.pages) {
-      for (const msg of page.messages) {
-        if (newest === undefined || msg.sequence > newest) {
-          newest = msg.sequence;
-        }
-      }
-    }
-    return newest;
-  }, [historyData]);
-
-  // Subscribe to new messages via SSE - update cache directly
-  // Pass cursor to catch up on missed messages when connecting/reconnecting
-  // Handles both complete messages (persisted) and partial messages (transient streaming updates)
-  trpc.sse.onNewMessage.useSubscription(
-    { sessionId, afterSequence: newestSequence },
-    {
-      onData: (trackedData) => {
-        const newMessage = trackedData.data.message;
-        const isPartial = newMessage.id.startsWith('partial-');
-
-        // Add message directly to the infinite query cache
-        utils.claude.getHistory.setInfiniteData({ sessionId, limit: MESSAGE_PAGE_SIZE }, (old) => {
-          if (!old) {
-            // No existing data - create initial page
-            return {
-              pages: [{ messages: [newMessage], hasMore: false }],
-              pageParams: [{ direction: 'backward' as const, sequence: undefined }],
-            };
-          }
-
-          if (isPartial) {
-            // Partial message: replace existing partial or add new one
-            const newPages = old.pages.map((page, pageIndex) => {
-              if (pageIndex !== 0) return page;
-              // Replace existing partial, or append if none exists
-              const hasExistingPartial = page.messages.some((m) => m.id.startsWith('partial-'));
-              if (hasExistingPartial) {
-                return {
-                  ...page,
-                  messages: page.messages.map((m) =>
-                    m.id.startsWith('partial-') ? newMessage : m
-                  ),
-                };
-              }
-              return { ...page, messages: [...page.messages, newMessage] };
-            });
-            return { ...old, pages: newPages };
-          }
-
-          // Complete message: remove any partial messages and add the new one
-          // Check for deduplication first
-          for (const page of old.pages) {
-            if (page.messages.some((m) => m.id === newMessage.id)) {
-              return old; // Already have this message
-            }
-          }
-
-          const newPages = [...old.pages];
-          // Remove partial messages from the first page (they're replaced by complete messages)
-          const firstPageMessages = newPages[0].messages.filter(
-            (m) => !m.id.startsWith('partial-')
-          );
-          newPages[0] = {
-            ...newPages[0],
-            messages: [...firstPageMessages, newMessage],
-          };
-
-          return { ...old, pages: newPages };
-        });
-
-        // Refetch token usage only for complete messages (partials don't affect totals)
-        if (!isPartial) {
-          refetchTokenUsage();
-        }
-      },
-      onError: (err) => {
-        console.error('Message SSE error:', err);
-      },
-    }
-  );
-
-  // Flatten bidirectional pages into chronological order
-  // Pages array structure:
-  // - pages[0] = newest (from fetchPreviousPage, or initial if no previous fetched)
-  // - pages[n-1] = oldest (from fetchNextPage)
-  // Each page's messages are already in chronological order
-  const messages = useMemo(() => {
-    if (!historyData?.pages) return [];
-
-    const result: Message[] = [];
-    // Reverse pages to get oldest-first, then flatten
-    for (const page of [...historyData.pages].reverse()) {
-      for (const msg of page.messages) {
-        result.push({
-          id: msg.id,
-          type: msg.type,
-          content: msg.content,
-          sequence: msg.sequence,
-        });
-      }
-    }
-    return result;
-  }, [historyData]);
-
-  return {
-    messages,
-    isLoading,
-    isFetchingMore: isFetchingNextPage,
-    hasMore: hasNextPage ?? false,
-    fetchMore: fetchNextPage,
-    tokenUsage: tokenUsageData,
-  };
-}
-
-/**
- * Hook for showing a notification when Claude finishes processing.
- * Only shows notification if the page was hidden while Claude was working.
- */
-function useWorkCompleteNotification(
-  sessionName: string | undefined,
-  isWorking: boolean,
-  showNotification: (title: string, options?: NotificationOptions) => Promise<void>
-) {
-  const wasWorkingRef = useRef(false);
-  const wasHiddenWhileWorkingRef = useRef(false);
-
-  // Track when we become hidden while working
-  useEffect(() => {
-    const handleVisibilityChange = () => {
-      if (document.hidden && isWorking) {
-        wasHiddenWhileWorkingRef.current = true;
-      }
-    };
-
-    document.addEventListener('visibilitychange', handleVisibilityChange);
-
-    // Also check initial state - if we're hidden and working, mark it
-    if (document.hidden && isWorking) {
-      wasHiddenWhileWorkingRef.current = true;
-    }
-
-    return () => {
-      document.removeEventListener('visibilitychange', handleVisibilityChange);
-    };
-  }, [isWorking]);
-
-  // Detect transition from working to not working
-  useEffect(() => {
-    // When work completes (was working -> not working)
-    if (wasWorkingRef.current && !isWorking && wasHiddenWhileWorkingRef.current) {
-      // Only notify if we're still hidden (user hasn't come back yet)
-      if (document.hidden && sessionName) {
-        showNotification('Claude finished', {
-          body: `Work complete on ${sessionName}`,
-          tag: 'work-complete',
-        });
-      }
-      wasHiddenWhileWorkingRef.current = false;
-    }
-
-    wasWorkingRef.current = isWorking;
-  }, [isWorking, sessionName, showNotification]);
-}
-
-/**
- * Hook for managing Claude process state: running status, send prompts, interrupt, and commands.
- */
-function useClaudeState(sessionId: string) {
-  const utils = trpc.useUtils();
-
-  // Fetch Claude running state
-  const { data: runningData, refetch } = trpc.claude.isRunning.useQuery({ sessionId });
-
-  // Fetch available slash commands
-  const { data: commandsData, refetch: refetchCommands } = trpc.claude.getCommands.useQuery(
-    { sessionId },
-    { staleTime: Infinity }
-  );
-
-  // Refetch when app regains visibility or network reconnects
-  const refetchAll = useCallback(() => {
-    refetch();
-    refetchCommands();
-  }, [refetch, refetchCommands]);
-  useRefetchOnReconnect(refetchAll);
-
-  // Subscribe to Claude running state via SSE - update cache directly
-  trpc.sse.onClaudeRunning.useSubscription(
-    { sessionId },
-    {
-      onData: (trackedData) => {
-        utils.claude.isRunning.setData({ sessionId }, { running: trackedData.data.running });
-        // When Claude finishes running, refetch commands in case new ones were discovered
-        if (!trackedData.data.running) {
-          refetchCommands();
-        }
-      },
-      onError: (err) => {
-        console.error('Claude running SSE error:', err);
-      },
-    }
-  );
-
-  // Subscribe to commands updates via SSE - update cache directly
-  trpc.sse.onCommands.useSubscription(
-    { sessionId },
-    {
-      onData: (trackedData) => {
-        utils.claude.getCommands.setData({ sessionId }, { commands: trackedData.data.commands });
-      },
-      onError: (err) => {
-        console.error('Commands SSE error:', err);
-      },
-    }
-  );
-
-  const sendMutation = trpc.claude.send.useMutation();
-  const interruptMutation = trpc.claude.interrupt.useMutation();
-
-  const send = useCallback(
-    (prompt: string) => {
-      sendMutation.mutate({ sessionId, prompt });
-    },
-    [sessionId, sendMutation]
-  );
-
-  const interrupt = useCallback(() => {
-    interruptMutation.mutate({ sessionId });
-  }, [sessionId, interruptMutation]);
-
-  const isRunning = runningData?.running ?? false;
-  const commands = commandsData?.commands ?? [];
-
-  return {
-    isRunning,
-    send,
-    interrupt,
-    isInterrupting: interruptMutation.isPending,
-    commands,
-  };
-}
+import { useSessionState } from '@/hooks/useSessionState';
+import { useSessionMessages } from '@/hooks/useSessionMessages';
+import { useClaudeState } from '@/hooks/useClaudeState';
+import { useWorkCompleteNotification } from '@/hooks/useWorkCompleteNotification';
 
 function SessionView({ sessionId }: { sessionId: string }) {
   // Session state: data, start/stop/archive
