+++ src/hooks/useInfiniteScroll.ts
@@ -1,0 +1,44 @@
+import { useEffect, useRef } from 'react';
+
+interface UseInfiniteScrollOptions {
+  hasNextPage: boolean;
+  isFetchingNextPage: boolean;
+  fetchNextPage: () => void;
+  rootMargin?: string;
+}
+
+/**
+ * Hook that uses IntersectionObserver to trigger infinite scroll pagination
+ * when a sentinel element becomes visible within a scroll container.
+ *
+ * Returns refs to attach to the scroll container and the sentinel element.
+ */
+export function useInfiniteScroll({
+  hasNextPage,
+  isFetchingNextPage,
+  fetchNextPage,
+  rootMargin = '100px',
+}: UseInfiniteScrollOptions) {
+  const scrollRef = useRef<HTMLDivElement>(null);
+  const sentinelRef = useRef<HTMLLIElement>(null);
+
+  useEffect(() => {
+    const sentinel = sentinelRef.current;
+    const scrollContainer = scrollRef.current;
+    if (!sentinel || !scrollContainer) return;
+
+    const observer = new IntersectionObserver(
+      (entries) => {
+        if (entries[0].isIntersecting && hasNextPage && !isFetchingNextPage) {
+          fetchNextPage();
+        }
+      },
+      { root: scrollContainer, rootMargin }
+    );
+
+    observer.observe(sentinel);
+    return () => observer.disconnect();
+  }, [hasNextPage, isFetchingNextPage, fetchNextPage, rootMargin]);
+
+  return { scrollRef, sentinelRef };
+}
